+++ ui/src/components/management/RouteInfoGrid.tsx
@@ -61,24 +61,29 @@ const RouteInfoGrid = ({
     const selectedRoute = permitRoutes.find((route) => route.id === routeId);
     if (selectedRoute) {
       setSelectedRouteOption(selectedRoute);
+      const { nextAvailableTransportNumber, routeBridges = [] } = selectedRoute;
 
       if (modifiedRouteTransportDetail && (!routeTransportId || routeTransportId === 0)) {
-        // This is a new route transport, so make sure supervision details are available for BridgeGrid
-        const { routeBridges = [] } = selectedRoute || {};
-        const newSupervisions: ISupervision[] = routeBridges.map((routeBridge) => {
-          const { id: routeBridgeId, contractBusinessId = "" } = routeBridge;
-          return {
-            id: 0,
-            routeBridgeId,
-            routeTransportId,
-            plannedTime: plannedDepartureTime ? moment(plannedDepartureTime).toDate() : moment().toDate(),
-            conformsToPermit: false,
-            // If contractBusinessId is provided from LeLu, supervisor is the area contractor. Otherwise, it's the current company.
-            supervisorCompany: contractBusinessId ? contractBusinessId : businessId,
-            supervisorType: contractBusinessId ? SupervisorType.AREA_CONTRACTOR : SupervisorType.OWN_SUPERVISOR,
-            routeBridge: routeBridge,
-          };
-        });
+        // This is a new route transport, so make sure supervision details are available for BridgeSupervisionGrid
+        let newSupervisions: ISupervision[] = [];
+
+        // Set supervisions only if there are valid transport numbers available for transport
+        if (nextAvailableTransportNumber) {
+          newSupervisions = routeBridges.map((routeBridge) => {
+            const { id: routeBridgeId, contractBusinessId = "" } = routeBridge;
+            return {
+              id: 0,
+              routeBridgeId,
+              routeTransportId,
+              plannedTime: plannedDepartureTime ? moment(plannedDepartureTime).toDate() : moment().toDate(),
+              conformsToPermit: false,
+              // If contractBusinessId is provided from LeLu, supervisor is the area contractor. Otherwise, it's the current company.
+              supervisorCompany: contractBusinessId ? contractBusinessId : businessId,
+              supervisorType: contractBusinessId ? SupervisorType.AREA_CONTRACTOR : SupervisorType.OWN_SUPERVISOR,
+              routeBridge: routeBridge,
+            };
+          });
+        }
         const newDetail: IRouteTransport = {
           ...modifiedRouteTransportDetail,
           routeId,
@@ -125,6 +130,7 @@ const RouteInfoGrid = ({
                       })}
                     selectedValue={selectedRouteId}
                     onChange={(routeId) => selectRoute(routeId as number)}
+                    disabled={!!routeTransportId} // Prevent changing route when transport has been saved
                   />
                 ) : (
                   <IonText>{selectedRouteName}</IonText>
